Parametrise reservation date with inline strings
This commit mostly serves as an example to illustrate that for some
types, a hack to parametrise them is available. You can only use
primitive types like bool, int, string, etc. in attributes, so for other
types of data, [InlineData] seems to be incompatible.

If, however, there's a conversion from a primitive type to the type you
need, you can use the primitive type with [InlineData] and then convert
to the type that you need in the test method itself.

Here, I've chosen to encode dates as strings and then subsequently parse
them, but I could also have chosen to encode dates as longs and used the
DateTime constructor overload that takes a number of ticks to initialise
the value. It's easier, however, to relate to a value like "2018-08-30"
than the value 636711840000000000, which is the reason I chose strings
for this example.