Log from Post method
This change is intended to provide a condensed example of the way
logging is often done in many enterprise code bases. There's so many
issues with this style of logging that I hardly know where to begin, but
here goes:

Lots of logging, but not enough detail. A message like "Entering Post
method..." is of little use when not accompanied with details about
method arguments and (possibly) program state.

Ambiguous logging. Two log statements log the same message ("Invalid
reservation date."), which means that it may not be clear from the log
entry alone which of the lines of code that can produce this entry
actually did so. You have to look at the previous log entries to piece
together what happened. If you see a previous log entry that states
"Mapping DTO to Domain Model." you know that the reservation date was
invalid because it was in the past. If you don't see that previous log
entry, you know that the reservation date was malformed.

No particular consistency when it comes to the use of log levels (debug,
info, warning, error (not shown)). For instance, there's no reason to
log the return of 500 Internal Server Error as a warning when this is an
entirely expected outcome.